+++ CHANGELOG.md
@@ -6,7 +6,23 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#23](https://github.com/zendframework/zend-expressive-zendviewrenderer/pull/23)
+  adds the ability to disable layouts either globally or when rendering. Disable
+  globally by setting the default `layout` parameter to boolean `false`:
+
+  ```php
+  $renderer->addDefaultParam(TemplateRendererInterface::TEMPLATE_ALL, 'layout', false);
+  ```
+
+  Or do so when rendering, by passing the template variable `layout` with a
+  boolean `false` value:
+
+  ```php
+  $renderer->render($templateName, [
+      'layout' => false,
+      // other template variables
+  ]);
+  ```
 
 ### Deprecated